Split controller.go off of the Route reconciler
