+++ pkg/reconciler/route/route.go
@@ -27,27 +27,21 @@ import (
 	"k8s.io/apimachinery/pkg/runtime/schema"
 	"k8s.io/apimachinery/pkg/types"
 	"k8s.io/apimachinery/pkg/util/sets"
-	corev1informers "k8s.io/client-go/informers/core/v1"
 	corev1listers "k8s.io/client-go/listers/core/v1"
 	"k8s.io/client-go/tools/cache"
 
 	"github.com/knative/pkg/apis"
 	duckv1alpha1 "github.com/knative/pkg/apis/duck/v1alpha1"
 	duckv1beta1 "github.com/knative/pkg/apis/duck/v1beta1"
-	"github.com/knative/pkg/configmap"
 	"github.com/knative/pkg/controller"
 	"github.com/knative/pkg/logging"
 	"github.com/knative/pkg/system"
 	"github.com/knative/pkg/tracker"
 	"github.com/knative/serving/pkg/apis/networking"
 	netv1alpha1 "github.com/knative/serving/pkg/apis/networking/v1alpha1"
-	"github.com/knative/serving/pkg/apis/serving"
 	"github.com/knative/serving/pkg/apis/serving/v1alpha1"
-	networkinginformers "github.com/knative/serving/pkg/client/informers/externalversions/networking/v1alpha1"
-	servinginformers "github.com/knative/serving/pkg/client/informers/externalversions/serving/v1alpha1"
 	networkinglisters "github.com/knative/serving/pkg/client/listers/networking/v1alpha1"
 	listers "github.com/knative/serving/pkg/client/listers/serving/v1alpha1"
-	"github.com/knative/serving/pkg/network"
 	"github.com/knative/serving/pkg/reconciler"
 	"github.com/knative/serving/pkg/reconciler/route/config"
 	"github.com/knative/serving/pkg/reconciler/route/domains"
@@ -55,10 +49,6 @@ import (
 	resourcenames "github.com/knative/serving/pkg/reconciler/route/resources/names"
 	"github.com/knative/serving/pkg/reconciler/route/traffic"
 	tr "github.com/knative/serving/pkg/reconciler/route/traffic"
-)
-
-const (
-	controllerAgentName = "route-controller"
 )
 
 // routeFinalizer is the name that we put into the resource finalizer list, e.g.
@@ -69,11 +59,6 @@ var (
 	routeResource  = v1alpha1.Resource("routes")
 	routeFinalizer = routeResource.String()
 )
-
-type configStore interface {
-	ToContext(ctx context.Context) context.Context
-	WatchConfigs(w configmap.Watcher)
-}
 
 // Reconciler implements controller.Reconciler for Route resources.
 type Reconciler struct {
@@ -94,105 +79,6 @@ type Reconciler struct {
 
 // Check that our Reconciler implements controller.Reconciler
 var _ controller.Reconciler = (*Reconciler)(nil)
-
-// NewController initializes the controller and is called by the generated code
-// Registers eventhandlers to enqueue events
-// config - client configuration for talking to the apiserver
-// si - informer factory shared across all controllers for listening to events and indexing resource properties
-// reconcileKey - function for mapping queue keys to resource names
-func NewController(
-	opt reconciler.Options,
-	routeInformer servinginformers.RouteInformer,
-	configInformer servinginformers.ConfigurationInformer,
-	revisionInformer servinginformers.RevisionInformer,
-	serviceInformer corev1informers.ServiceInformer,
-	clusterIngressInformer networkinginformers.ClusterIngressInformer,
-	certificateInformer networkinginformers.CertificateInformer,
-) *controller.Impl {
-	return NewControllerWithClock(opt, routeInformer, configInformer, revisionInformer,
-		serviceInformer, clusterIngressInformer, certificateInformer, system.RealClock{})
-}
-
-func NewControllerWithClock(
-	opt reconciler.Options,
-	routeInformer servinginformers.RouteInformer,
-	configInformer servinginformers.ConfigurationInformer,
-	revisionInformer servinginformers.RevisionInformer,
-	serviceInformer corev1informers.ServiceInformer,
-	clusterIngressInformer networkinginformers.ClusterIngressInformer,
-	certificateInformer networkinginformers.CertificateInformer,
-	clock system.Clock,
-) *controller.Impl {
-
-	// No need to lock domainConfigMutex yet since the informers that can modify
-	// domainConfig haven't started yet.
-	c := &Reconciler{
-		Base:                 reconciler.NewBase(opt, controllerAgentName),
-		routeLister:          routeInformer.Lister(),
-		configurationLister:  configInformer.Lister(),
-		revisionLister:       revisionInformer.Lister(),
-		serviceLister:        serviceInformer.Lister(),
-		clusterIngressLister: clusterIngressInformer.Lister(),
-		certificateLister:    certificateInformer.Lister(),
-		clock:                clock,
-	}
-	impl := controller.NewImpl(c, c.Logger, "Routes", reconciler.MustNewStatsReporter("Routes", c.Logger))
-
-	c.Logger.Info("Setting up event handlers")
-	routeInformer.Informer().AddEventHandler(reconciler.Handler(impl.Enqueue))
-
-	serviceInformer.Informer().AddEventHandler(cache.FilteringResourceEventHandler{
-		FilterFunc: controller.Filter(v1alpha1.SchemeGroupVersion.WithKind("Route")),
-		Handler:    reconciler.Handler(impl.EnqueueControllerOf),
-	})
-
-	clusterIngressInformer.Informer().AddEventHandler(reconciler.Handler(
-		impl.EnqueueLabelOfNamespaceScopedResource(
-			serving.RouteNamespaceLabelKey, serving.RouteLabelKey)))
-
-	c.tracker = tracker.New(impl.EnqueueKey, opt.GetTrackerLease())
-
-	configInformer.Informer().AddEventHandler(reconciler.Handler(
-		// Call the tracker's OnChanged method, but we've seen the objects
-		// coming through this path missing TypeMeta, so ensure it is properly
-		// populated.
-		controller.EnsureTypeMeta(
-			c.tracker.OnChanged,
-			v1alpha1.SchemeGroupVersion.WithKind("Configuration"),
-		),
-	))
-
-	revisionInformer.Informer().AddEventHandler(reconciler.Handler(
-		// Call the tracker's OnChanged method, but we've seen the objects
-		// coming through this path missing TypeMeta, so ensure it is properly
-		// populated.
-		controller.EnsureTypeMeta(
-			c.tracker.OnChanged,
-			v1alpha1.SchemeGroupVersion.WithKind("Revision"),
-		),
-	))
-
-	certificateInformer.Informer().AddEventHandler(cache.FilteringResourceEventHandler{
-		FilterFunc: controller.Filter(v1alpha1.SchemeGroupVersion.WithKind("Route")),
-		Handler:    reconciler.Handler(impl.EnqueueControllerOf),
-	})
-
-	c.Logger.Info("Setting up ConfigMap receivers")
-	configsToResync := []interface{}{
-		&network.Config{},
-		&config.Domain{},
-	}
-	resync := configmap.TypeFilter(configsToResync...)(func(string, interface{}) {
-		impl.GlobalResync(routeInformer.Informer())
-	})
-	c.configStore = config.NewStore(c.Logger.Named("config-store"), resync)
-	c.configStore.WatchConfigs(opt.ConfigMapWatcher)
-	return impl
-}
-
-/////////////////////////////////////////
-//  Event handlers
-/////////////////////////////////////////
 
 // Reconcile compares the actual state with the desired, and attempts to
 // converge the two. It then updates the Status block of the Route resource
